Add virtual IRQ for serial receive
We want to keep the UART driver in the kernel to ensure that reset
signals are caught. However, we also want to be able to access serial
input from userspace (for GDB). This patch adds a shared queue between
userspace/kernel for the received characters and a virtual interrupt to
allow userspace to wait for characters without the actual UART IRQ.

Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>